requirements.txt
anyio==4.3.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
bcc==0.1.10
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==6.1.0
blinker==1.7.0
Brlapi>=0.8.3
Brotli==1.0.9
bs4==0.0.2
cachetools==5.3.3
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
chrome-gnome-shell==0.0.0
click==8.1.7
colorama==0.4.4
comm==0.2.1
command-not-found==0.3
contourpy==1.2.0
cppman==0.5.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
debugpy==1.8.1
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
Flask==3.0.1
Flask-WTF==1.2.1
fonttools==4.49.0
fqdn==1.5.1
fs==2.4.12
gast==0.5.2
google-auth==2.28.2
google-auth-oauthlib==1.2.0
gspread==6.0.2
gunicorn==21.2.0
h11==0.14.0
hidpidaemon==18.4.6
html5lib==1.1
httpcore==1.0.4
httplib2==0.20.2
httpx==0.27.0
idna==3.3
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib-metadata==4.6.4
install==1.3.5
ipykernel==6.29.2
ipython==8.22.1
ipython_genutils==0.2.0
ipywidgets==6.0.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.3
json5==0.9.17
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.4.0
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
kernelstub==3.1.4
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
MarkupSafe==2.1.4
matplotlib==3.8.3
matplotlib-inline==0.1.6
mistune==3.0.2
more-itertools==8.10.0
moviepy==1.0.3
mpmath==0.0.0
mysql-connector-python==8.3.0
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nest-asyncio==1.6.0
netaddr==0.8.0
netifaces==0.11.0
notebook==6.4.8
notebook_shim==0.2.4
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
opencv-python==4.9.0.80
overrides==7.7.0
packaging==23.2
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.2.0
ply==3.11
pop-transition==1.1.2
proglog==0.1.10
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==3.12.4
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydbus==0.6.0
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMySQL==1.1.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-gnupg==0.4.8
python-json-logger==2.0.7
python-xlib==0.29
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.33.0
repolib==2.2.1
repoman==1.4.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
rsa==4.9
scipy==1.8.0
SecretStorage==3.3.1
Send2Trash==1.8.2
sessioninstaller==0.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.1
stack-data==0.6.3
StrEnum==0.4.15
sympy==1.9
systemd-python==234
terminado==0.18.0
testpath==0.5.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unicodedata2==14.0.0
uri-template==1.3.0
urllib3==1.26.5
wadllib==1.3.6
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==2.0.0
wordcloud==1.9.3
WTForms==3.1.2
xdg==5
xkit==0.0.0
zipp==1.0.0
